feat: move joinsound channel from config to /admin joinsounds
